requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlmodel==0.0.14
python-multipart==0.0.6
jinja2==3.1.3
itsdangerous==2.1.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
weasyprint==60.2
pytz==2024.1
aiosmtplib==3.0.1
aiofiles==23.2.1
Pillow==10.2.0
python-dotenv==1.0.0
apscheduler==3.10.4
pypdf==4.0.1
reportlab==4.0.7
fillpdf==0.7.3
PyMuPDF==1.24.9
# PostgreSQL driver
psycopg2-binary==2.9.9
# Cloudinary for PDF storage
cloudinary==1.36.0